pyproject.toml
[project]
name = "agent-dynamic-tasks"
version = "0.1.0"
description = "Agent coordination system with dynamic task breakdown and execution"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.1.0",
    "langgraph>=0.1.0", 
    "openai>=1.0.0",
    "python-dotenv>=1.0.0"
]
